Refactor Protocol folder in  cardano-config
